move all modules from utils to views, remove utils folder
