+++ contributors/views/contributors_views/contributor_issues.py
@@ -2,7 +2,7 @@ from django.views import generic
 
 from contributors.forms.forms import NameStatusFilterForm
 from contributors.models import Contribution
-from contributors.views.utils.mixins import TableSortSearchAndPaginationMixin
+from contributors.views.mixins import TableSortSearchAndPaginationMixin
 
 
 class ListView(TableSortSearchAndPaginationMixin, generic.ListView):
